feat(sync): send sync_complete event to amplitude
