+++ .travis.yml
@@ -4,6 +4,9 @@ node_js:
 branches:
   only:
     - master
+before_install:
+  - sudo apt-get update -qq
+  - sudo apt-get install libboost-all-dev
 script:
   - npm run lint
