+++ sync-cron.sh
@@ -2,4 +2,5 @@
 
 cd /home/ec2-user/sync-amplitude
 . ./.env
-/usr/local/bin/node sync
+/usr/local/bin/node sync-events
+/usr/local/bin/node sync-summary
